tests/wpt/web-platform-tests/svg/animations/animate-keySplines.html

<!doctype html>
<html>
<meta charset="utf-8">
<title>Testing correct parsing of keySplines.</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/SVGAnimationTestCase-testharness.js"></script>

<svg>
</svg>

<script>
var rootSVGElement = document.querySelector("svg");
var epsilon = 1.0;

// Setup test document
var rect = createSVGElement("rect");
rect.setAttribute("id", "rect");
rect.setAttribute("fill", "green");
rect.setAttribute("x", "0");
rect.setAttribute("y", "0");
rect.setAttribute("width", "100");
rect.setAttribute("height", "100");
rect.setAttribute("onclick", "executeTest()");

var animate = createSVGElement("animate");
animate.setAttribute("id", "animation");
animate.setAttribute("attributeName", "height");
animate.setAttribute("calcMode", "spline");
animate.setAttribute("keyTimes", " 0 ; 0.3333333 ; 0.666666; 1 ");
animate.setAttribute("keySplines", "  0 ,0  1 , 1  ;   0 0 , 1 ,    1  ;  .75 , 0 , 0 , .75 ; ");
animate.setAttribute("values", "200;167;111;0");
animate.setAttribute("begin", "0s");
animate.setAttribute("dur", "9s");
rect.appendChild(animate);
rootSVGElement.appendChild(rect);

// Setup animation test
function sample1() {
    // Check initial/end conditions
    assert_approx_equals(rect.height.animVal.value, 167, epsilon);
    assert_equals(rect.height.baseVal.value, 100);
}

function sample2() {
    // Check half-time conditions
    assert_approx_equals(rect.height.animVal.value, 111, epsilon);
    assert_equals(rect.height.baseVal.value, 100);
}

function sample3() {
    // Check just before-end conditions
    assert_approx_equals(rect.height.animVal.value, 0, epsilon);
    assert_equals(rect.height.baseVal.value, 100);
}

function sample4() {
    // Check end conditions
    assert_approx_equals(rect.height.animVal.value, 100, epsilon);
    assert_equals(rect.height.baseVal.value, 100);
}

smil_async_test((t) => {
    const expectedValues = [
        // [animationId, time, sampleCallback]
        ["animation", 3.0,   sample1],
        ["animation", 6.0,   sample2],
        ["animation", 8.999, sample3],
        ["animation", 9.001, sample4]
    ];

    runAnimationTest(t, expectedValues);
});

</script>
