Safari-609.1.11.1/LayoutTests/imported/w3c/web-platform-tests/svg/animations/animate-keySplines-expected.txt


FAIL Testing correct parsing of keySplines. assert_approx_equals: expected 167 +/- 1 but got 100
